requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.3.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
